Compensate discount to invoice lines

+++ internal/logic/subscription/handler/subscription_pending_update.go
@@ -77,7 +77,7 @@ func HandlePendingUpdatePaymentSuccess(ctx context.Context, sub *entity.Subscrip
 		return false, err
 	}
 
-	user := query.GetUserAccountById(ctx, uint64(sub.UserId))
+	user := query.GetUserAccountById(ctx, sub.UserId)
 	merchant := query.GetMerchantById(ctx, sub.MerchantId)
 
 	SubscriptionNewTimeline(ctx, invoice)
